tools: add cqfd for automation and improve README
Added CQFD to allow running test and coverage without having to install
dependencies on the host machine.
Added a jenkins test with coverage as text output.
Also add junit.xml output of pytest for jenkins build.
Added flake8 in requirements-dev as it was missing.
Added explanation of both testing methods in README.