frontend: add basic vite, react, tailwindcss setup
This commit adds a basic setup for a frontend using Vite, React and
TailwindCSS. Because this basic frontend need at least one page, also
include the loading page but without any loading logic.

.gitkeep is needed because it will copy to static/ folder in the
flask backend, and vite erase older content so we need it in front to
be copied in backend on build step.