bin: add vulnscout script to run docker scanner
This script is a utility for developers to run the docker in a more
convenient way. It is a wrapper which load config from files and run
the docker with the right parameters. It replace the old way to run
which was using a docker-compose file.
Reworked documentation for better readability and separation.