bin: add version check when running vulnscout.sh
Vulnscout try to reach Gerrit repository to list versions.
If the version is not found, it produce a warning message.
If a version is found and newer than the current one, it notice the user
A next release will implement the vulnscout update command.
Docker image tag is now bound to the vulnscout version.