front: link legend UX between charts in dashboard
On the Dashboard/Metrics page, default click action on legend item is to
toggle the visibility of this item in the chart. This commit make
hidding an item in some chart to also hide it in the other charts.

eg: if you hide the 'Fixed' status, the chart showing the severity
repartition will update to exclude all fixed vulnerabilities.

Note: Chart.JS interaction cannot be unit-tested.