front: pkg: fix bug when selecting all sources
On the package table, when using the filter to select only a source
(eg. yocto), and then re-change it to "All sources", the table was not
refreshed. Instead of setting filter to `undefined`, it was set on
"undefined" which result in no packages matching the filter.

Added test to avoid regression.