vuln: add EPSS scoring data from first.org
EPSS score is a pair of two metrics to estimate exploitability of vulns.
score is an index [0-1] of how likely a vuln is to be exploited.
percentile is an index [0-1] of how many vulns are less likely to be
exploited than this one. Convey a meaning of "this vuln is more likely
to be exploited than X% of all vulns".