template: add templating engine using jinja2
In order to be abled to generate reports and documents from our analysis
results, we need to implement a templating engine. We choose Jinja2 as
it powerfull, have security features and is well documented. Official
templates will be added in `src/views/templates` and custom one can be
added in `.vulnscout/templates` folder, directly in projects.
Documentation about available fields and filter is available as
`WRITING_TEMPLATES.adoc` file.