front: add optimization to replace O(n*n) by O(2n)
In frontend, enrichment functions (enrich vuln and enrich package) were
using .filter inside a .map, which lead to a O(n*n) complexity. This
commit extract the filter outside the loop, which reduce the complexity
to O(2n).
It reduce loading time between 300ms-1500ms on Yocto project with 6000
vulnerabilities.