views: add JSON file to save/load time estimates
Time estimation is a feature not supported by OpenVex, CycloneDX or SPDX
formats. In order to save and being able to restore time estimates,
we must save them in a file which will be loaded on next use.
None standard format for this king of tasks description was found, so
it is just a simple JSON format with vuln id + PERT time estimates.