api: add PATCH method to edit vulnerabilities
Add a PATCH method to allow user to edit and save time estimations for
a vulnerability. REST API should return the full object after
modification, so I also added a GET method as the code was already here.
It may be used  by thrid-party scripts to fetch informations about
a vulnerability alone with it's ID.