templates: add scan date when available
Add a global variable named `scan_date` in form `<date> at <time> (UTC)`
to indicated in report when the scan was done. This value should not be
editable manually, nor having a default value.

Because vulnscout can be run with `DEBUG_SKIP_SCAN=true`, scan date is
not garanteed to be available. In such case, it is essential to use
"unknown date" and not guessing from current date, as scan can be old.