spdx: add SPDX 3.0.1 support
Add a parser for SPDX 3.0.1 SBOMs generated by Yocto.

During the initial SPDX merge, only package data is extracted.
The existing implementation for SPDX 2.x files is then used
to generate merged.spdx.json, which is scanned by Grype.

In the final merge step, package, vulnerability, and assessment
data are extracted and combined with the Grype output to produce
the final report.