cleanup: Changed Dockerfile and provided example
.vulscount: Now use a folder "example" which is more explicit than "test
scan.sh: This script is used when the container starts. No reason to keep it at the root of the project placed in /src
Dockerfile: Modified to use scan.sh in /scan/src/scan.sh and not /scan/scan.sh as before
start-dev.sh: Modified to use example folder. Also, fixed issue with "vite" node module not closed correctly
.gitignore: Modified for example folder
docker: Now requirements.txt and requirements-dev.txt are placed in this folder to not pollute the root of the project