vulnscout.sh: Add script for manual VS usage
This script helps Vulnscout users to configure the tool for their SBOM.

It can be used to configure a YAML file and start a container.

In CI mode, it will perform a simple SBOM scan.

This script is not a replacement for the Yocto meta-vulnscout, but as
an alternative solution for projects that aren't Yocto-based.

Change-Id: Iec2c799e7de7a842576864c5be1bc35f2256bbfc