vulnscout_CI_test.sh: Add automatic test for CI mode
Create a vulnscout_CI_test.sh to test the CI mode with cqfd. It test if
the fail condition is correctly working and if the output files are
correctly created when the fail condition was not triggered.
To make it work, it was needed to install docker in the cqfd for DinD
context.